Refactor code and display time range
